test fix and refactor
