+++ dashboard/src/model/waiterModel.ts
@@ -1,8 +1,11 @@
 import {makeAutoObservable} from 'mobx';
 import {WaiterIDO} from '../../../api';
 
+export type assignedWaitersType = {orderID: string; waiterIds: string[]}[];
+
 export default class waiterModel {
 	_waiters: WaiterIDO[] = [];
+	_assignedWaiters: assignedWaitersType = [];
 
 	constructor() {
 		makeAutoObservable(this);
@@ -16,5 +19,29 @@ export default class waiterModel {
 	get waiters(): WaiterIDO[] {
 		console.info('Getting waiters', this._waiters);
 		return this._waiters;
+	}
+
+	get assignedWaiters(): assignedWaitersType {
+		return this._assignedWaiters;
+	}
+
+	set assignedWaiters(assignedWaiters: assignedWaitersType) {
+		console.info('Setting assigned waiters');
+		this._assignedWaiters = assignedWaiters;
+	}
+
+	updateAssignedWaiters(orderID: string, waiterIds: string[]): void {
+		console.info('Updating assigned waiters with ', orderID, waiterIds);
+		const assignedWaiterObject = this._assignedWaiters.find(
+			entry => entry.orderID === orderID
+		);
+		if (assignedWaiterObject !== undefined) {
+			assignedWaiterObject.waiterIds = waiterIds;
+		} else {
+			this._assignedWaiters.push({
+				orderID: orderID,
+				waiterIds: waiterIds,
+			});
+		}
 	}
 }
